requirements.txt
torch>=2.5.0
transformers>=4.51.0
peft>=0.15.0
datasets>=3.0.0
accelerate>=1.0.0
bitsandbytes>=0.45.0
gradio>=5.0.0
pyyaml>=6.0
sentencepiece>=0.2.0
protobuf>=4.25.0
